2025 한이음/Trial&Error/기타/news_trend_refactor/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
beautifulsoup4>=4.12
langchain-openai>=0.1.7
langchain-core>=0.1.53
python-dotenv>=1.0
pydantic>=2.6
rouge-score>=0.1.2
bert-score>=0.3.13
transformers>=4.41
torch>=2.1
regex>=2024.5.15
